refactor(server): rename types to schema for more more clear
